NDRS-701: Extract `State` to separate file.
