+++ node/src/components/linear_chain_sync/event.rs
@@ -1,6 +1,9 @@
 use crate::{
     components::fetcher::FetchResult,
-    types::{Block, BlockHash, BlockHeader},
+    effect::requests::{
+        BlockExecutorRequest, BlockValidationRequest, FetcherRequest, StorageRequest,
+    },
+    types::{Block, BlockByHeight, BlockHash, BlockHeader},
 };
 use std::fmt::{Debug, Display};
 
@@ -55,4 +58,24 @@ where
             }
         }
     }
+}
+
+pub trait ReactorEventT<I>:
+    From<StorageRequest>
+    + From<FetcherRequest<I, Block>>
+    + From<FetcherRequest<I, BlockByHeight>>
+    + From<BlockValidationRequest<BlockHeader, I>>
+    + From<BlockExecutorRequest>
+    + Send
+{
+}
+
+impl<I, REv> ReactorEventT<I> for REv where
+    REv: From<StorageRequest>
+        + From<FetcherRequest<I, Block>>
+        + From<FetcherRequest<I, BlockByHeight>>
+        + From<BlockValidationRequest<BlockHeader, I>>
+        + From<BlockExecutorRequest>
+        + Send
+{
 }
